Move functions from utils package to files package
