+++ files/markdown.go
@@ -1,4 +1,4 @@
-package utils
+package files
 
 import (
 	"github.com/gomarkdown/markdown"
