+++ files/markdown_test.go
@@ -1,4 +1,4 @@
-package utils
+package files
 
 import (
 	"testing"
